@@ -1,6 +1,7 @@
-module.exports = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
+import tailwind from 'tailwindcss'
+import tailwindConfig from './tailwind.config.js'
+import autoprefixer from 'autoprefixer'
+
+export default {
+  plugins:[tailwind(tailwindConfig),autoprefixer]
 }
